requirements_ai_ocr.txt
# Core data processing
pandas>=1.3.0
numpy>=1.20.0

# PDF processing
pytesseract>=0.3.8
pdf2image>=1.16.0
Pillow>=8.0.0
PyPDF2>=3.0.0

# Enhanced AI OCR engines
easyocr>=1.6.0
google-cloud-vision>=3.0.0

# Date parsing
python-dateutil>=2.8.0

# System dependencies note:
# For pytesseract to work, you need to install Tesseract OCR:
# - macOS: brew install tesseract
# - Ubuntu/Debian: sudo apt-get install tesseract-ocr
# - Windows: Download from https://github.com/UB-Mannheim/tesseract/wiki

# For pdf2image to work, you need to install poppler:
# - macOS: brew install poppler
# - Ubuntu/Debian: sudo apt-get install poppler-utils
# - Windows: Download from https://blog.alivate.com.au/poppler-windows/

# For Google Vision API:
# 1. Create a Google Cloud project and enable Vision API
# 2. Create service account and download JSON key
# 3. Set environment variable: export GOOGLE_APPLICATION_CREDENTIALS="path/to/key.json"

# Optional: For GPU acceleration with EasyOCR
# torch>=1.9.0
# torchvision>=0.10.0
